Split compose.go into command-focussed go files
