+++ local/compose/compose.go
@@ -1,0 +1,66 @@
+/*
+   Copyright 2020 Docker Compose CLI authors
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package compose
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"strings"
+
+	"github.com/docker/compose-cli/api/compose"
+
+	"github.com/compose-spec/compose-go/types"
+	errdefs2 "github.com/docker/compose-cli/errdefs"
+	moby "github.com/docker/docker/api/types"
+	"github.com/docker/docker/client"
+	"github.com/sanathkr/go-yaml"
+)
+
+// NewComposeService create a local implementation of the compose.Service API
+func NewComposeService(apiClient *client.Client) compose.Service {
+	return &composeService{apiClient: apiClient}
+}
+
+type composeService struct {
+	apiClient *client.Client
+}
+
+func (s *composeService) Up(ctx context.Context, project *types.Project, detach bool) error {
+	return errdefs2.ErrNotImplemented
+}
+
+func getContainerName(c moby.Container) string {
+	// Names return container canonical name /foo  + link aliases /linked_by/foo
+	for _, name := range c.Names {
+		if strings.LastIndex(name, "/") == 0 {
+			return name[1:]
+		}
+	}
+	return c.Names[0][1:]
+}
+
+func (s *composeService) Convert(ctx context.Context, project *types.Project, format string) ([]byte, error) {
+	switch format {
+	case "json":
+		return json.MarshalIndent(project, "", "  ")
+	case "yaml":
+		return yaml.Marshal(project)
+	default:
+		return nil, fmt.Errorf("unsupported format %q", format)
+	}
+}
